+++ app/src/main/java/com/numero/material_gallery/components/appbar/top/CollapsingActivity.kt
@@ -7,7 +7,7 @@ import android.view.Menu
 import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
 import com.numero.material_gallery.R
-import com.numero.material_gallery.fragment.ThemeInfoBottomSheetDialog
+import com.numero.material_gallery.themeinfo.ThemeInfoBottomSheetDialog
 import com.numero.material_gallery.repository.ConfigRepository
 import kotlinx.android.synthetic.main.activity_collapsing.*
 import org.koin.android.ext.android.inject
